requirements.txt
asgiref==3.8.1
Django==4.2.20
sqlparse==0.5.3
typing_extensions==4.13.2
openai==1.12.0
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
